Rearrange classes into feature-based packages
