+++ signer-cloud-server/src/main/java/com/wultra/signercloud/server/document/DocumentContentRepository.java
@@ -15,9 +15,8 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.signercloud.server.repository;
+package com.wultra.signercloud.server.document;
 
-import com.wultra.signercloud.server.dao.DocumentContent;
 import org.springframework.data.repository.CrudRepository;
 
 /**
