+++ signer-cloud-server/src/main/java/com/wultra/signercloud/server/signer/Signer.java
@@ -15,9 +15,8 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.signercloud.server.dao;
+package com.wultra.signercloud.server.signer;
 
-import com.wultra.signercloud.server.status.SignerStatus;
 import lombok.Builder;
 import lombok.Getter;
 import org.springframework.data.annotation.Id;
